// src/generated/user/resources/user_pb.ts
// @generated by protoc-gen-es v2.6.0 with parameter "target=ts,import_extension=none"
// @generated from file string_phone/user/resources/user.proto (package string_phone.user.resources, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file string_phone/user/resources/user.proto.
 */
export const file_string_phone_user_resources_user: GenFile = /*@__PURE__*/
  fileDesc("CiZzdHJpbmdfcGhvbmUvdXNlci9yZXNvdXJjZXMvdXNlci5wcm90bxIbc3RyaW5nX3Bob25lLnVzZXIucmVzb3VyY2VzIqYBCgRVc2VyEgoKAmlkGAEgASgJEgwKBG5hbWUYAiABKAkSDgoGY3JlZGl0GAMgASgFEhQKDHBob25lX251bWJlchgEIAEoCRIuCgpjcmVhdGVkX2F0GAUgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcBIuCgp1cGRhdGVkX2F0GAYgASgLMhouZ29vZ2xlLnByb3RvYnVmLlRpbWVzdGFtcEI7WjlnaXRodWIuY29tL3N0cmluZy1waG9uZS9wcm90by9zdHJpbmdfcGhvbmUvdXNlci9yZXNvdXJjZXNiBnByb3RvMw", [file_google_protobuf_timestamp]);

/**
 * @generated from message string_phone.user.resources.User
 */
export type User = Message<"string_phone.user.resources.User"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: int32 credit = 3;
   */
  credit: number;

  /**
   * @generated from field: string phone_number = 4;
   */
  phoneNumber: string;

  /**
   * @generated from field: google.protobuf.Timestamp created_at = 5;
   */
  createdAt?: Timestamp;

  /**
   * @generated from field: google.protobuf.Timestamp updated_at = 6;
   */
  updatedAt?: Timestamp;
};

/**
 * Describes the message string_phone.user.resources.User.
 * Use `create(UserSchema)` to create a new message.
 */
export const UserSchema: GenMessage<User> = /*@__PURE__*/
  messageDesc(file_string_phone_user_resources_user, 0);